feat: refactor folder structure
