+++ src/App/store.js
@@ -1,7 +1,7 @@
 import { combineReducers, configureStore } from '@reduxjs/toolkit'
 import { persistReducer } from 'redux-persist'
 import storage from 'redux-persist/lib/storage' // defaults to localStorage for web
-import patientQueueReducer from '../PatientQueue/reducer'
+import patientQueueReducer from '../Patients/reducer'
  
 const persistConfig = {
   key: 'root',
