--- src/Map/Map.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Map = () => {
-  return (
-    <main>
-      <p>Map Goes Here.</p>
-    </main>
-  );
-}
-
-export default Map;
