+++ src/Queue/ReassignmentModal.jsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import Box from '@mui/material/Box';
-import VisitForm from "../VisitForm/VisitForm"
+import VisitForm from "../Visit/VisitForm"
 import Modal from '@mui/material/Modal';
 import Button from '@mui/material/Button';
 
@@ -18,7 +18,7 @@ const style = {
   p: 4,
 };
 
-const VisitsModal = () => {
+const ReassignmentModal = () => {
   const [open, setOpen] = useState(false);
   const handleOpen = () => setOpen(true);
   const handleClose = () => setOpen(false);
@@ -30,7 +30,7 @@ return(
       onClick={ handleOpen }
       size="small"
       variant="contained"
-    >assign
+    >reassign
     </Button>
     <Modal
       open={ open }
@@ -45,4 +45,4 @@ return(
   </>
 )
 }
-export default VisitsModal;
+export default ReassignmentModal;
